refactored and fixed tests
