+++ test/lib/tasks/bundlers/generateLibraryPreload.js
@@ -1,33 +1,6 @@
 const test = require("ava");
-const path = require("path");
-const chai = require("chai");
-chai.use(require("chai-fs"));
-const fs = require("graceful-fs");
-const {promisify} = require("util");
-const readFile = promisify(fs.readFile);
-const assert = chai.assert;
 const sinon = require("sinon");
 const mock = require("mock-require");
-
-const ui5Builder = require("../../../../");
-const builder = ui5Builder.builder;
-const libraryDPath = path.join(__dirname, "..", "..", "..", "fixtures", "library.d");
-const libraryNPath = path.join(__dirname, "..", "..", "..", "fixtures", "library.n");
-const sapUiCorePath = path.join(__dirname, "..", "..", "..", "fixtures", "sap.ui.core");
-
-const recursive = require("recursive-readdir");
-
-const findFiles = (folder) => {
-	return new Promise((resolve, reject) => {
-		recursive(folder, (err, files) => {
-			if (err) {
-				reject(err);
-			} else {
-				resolve(files);
-			}
-		});
-	});
-};
 
 test.beforeEach((t) => {
 	t.context.workspace = {
@@ -356,7 +329,7 @@ test("generateLibraryPreload for sap.ui.core (w/o ui5loader.js)", async (t) => {
 							"sap/ui/util/",
 							"sap/ui/Global.js",
 							"sap/ui/thirdparty/crossroads.js",
-							"sap/ui/thirdparty/caja-htmlsanitizer.js",
+							"sap/ui/thirdparty/caja-html-sanitizer.js",
 							"sap/ui/thirdparty/hasher.js",
 							"sap/ui/thirdparty/signals.js",
 							"sap/ui/thirdparty/jquery-mobile-custom.js",
@@ -623,7 +596,7 @@ test("generateLibraryPreload for sap.ui.core (/w ui5loader.js)", async (t) => {
 							"sap/ui/util/",
 							"sap/ui/Global.js",
 							"sap/ui/thirdparty/crossroads.js",
-							"sap/ui/thirdparty/caja-htmlsanitizer.js",
+							"sap/ui/thirdparty/caja-html-sanitizer.js",
 							"sap/ui/thirdparty/hasher.js",
 							"sap/ui/thirdparty/signals.js",
 							"sap/ui/thirdparty/jquery-mobile-custom.js",
@@ -660,186 +633,3 @@ test("generateLibraryPreload for sap.ui.core (/w ui5loader.js)", async (t) => {
 	t.true(ReaderCollectionPrioritizedStub.calledWithNew(),
 		"ReaderCollectionPrioritized should have been called with 'new'");
 });
-
-
-test("integration: build library.d with library preload", async (t) => {
-	const destPath = "./test/tmp/build/library.d/preload";
-	const expectedPath = "./test/expected/build/library.d/preload";
-	const excludedTasks = ["*"];
-	const includedTasks = ["generateLibraryPreload"];
-
-	return t.notThrowsAsync(builder.build({
-		tree: libraryDTree,
-		destPath,
-		excludedTasks,
-		includedTasks
-	}).then(() => {
-		return findFiles(expectedPath);
-	}).then((expectedFiles) => {
-		// Check for all directories and files
-		assert.directoryDeepEqual(destPath, expectedPath);
-
-		// Check for all file contents
-		t.deepEqual(expectedFiles.length, 4, "4 files are expected");
-		expectedFiles.forEach((expectedFile) => {
-			const relativeFile = path.relative(expectedPath, expectedFile);
-			const destFile = path.join(destPath, relativeFile);
-			assert.fileEqual(destFile, expectedFile);
-		});
-	}));
-});
-
-const libraryDTree = {
-	"id": "library.d",
-	"version": "1.0.0",
-	"path": libraryDPath,
-	"dependencies": [],
-	"_level": 1,
-	"specVersion": "0.1",
-	"type": "library",
-	"metadata": {
-		"name": "library.d",
-		"copyright": "Some fancy copyright"
-	},
-	"resources": {
-		"configuration": {
-			"paths": {
-				"src": "main/src",
-				"test": "main/test"
-			}
-		},
-		"pathMappings": {
-			"/resources/": "main/src",
-			"/test-resources/": "main/test"
-		}
-	}
-};
-
-test("integration: build library.n with library preload", (t) => {
-	const destPath = "./test/tmp/build/library.n/preload";
-	const expectedPath = "./test/expected/build/library.n/preload";
-	const excludedTasks = ["*"];
-	const includedTasks = ["generateLibraryPreload", "generateResourcesJson"];
-
-	return builder.build({
-		tree: libraryNTree,
-		destPath,
-		excludedTasks,
-		includedTasks
-	}).then(() => {
-		return findFiles(expectedPath);
-	}).then((expectedFiles) => {
-		// Check for all directories and files
-		assert.directoryDeepEqual(destPath, expectedPath);
-
-		// Check for all file contents
-		return checkFileContentsIgnoreLineFeeds(t, expectedFiles, expectedPath, destPath);
-	}).then(() => {
-		t.pass();
-	});
-});
-
-const libraryNTree = {
-	"id": "library.n",
-	"version": "1.0.0",
-	"path": libraryNPath,
-	"dependencies": [],
-	"_level": 0,
-	"specVersion": "0.1",
-	"type": "library",
-	"metadata": {
-		"name": "library.n",
-		"namespace": "library/n",
-		"copyright": "UI development toolkit for HTML5 (OpenUI5)\n * (c) Copyright 2009-xxx SAP SE or an SAP affiliate company.\n * Licensed under the Apache License, Version 2.0 - see LICENSE.txt."
-	},
-	"resources": {
-		"configuration": {
-			"paths": {
-				"src": "src",
-				"test": "test"
-			},
-			"propertiesFileSourceEncoding": "ISO-8859-1"
-		},
-		"pathMappings": {
-			"/resources/": "src",
-			"/test-resources/": "test"
-		}
-	}
-};
-
-
-test("integration: build sap.ui.core with library preload", async (t) => {
-	const destPath = "./test/tmp/build/sap.ui.core/preload";
-	const expectedPath = "./test/expected/build/sap.ui.core/preload";
-	const excludedTasks = ["*"];
-	const includedTasks = ["generateLibraryPreload"];
-
-	return t.notThrowsAsync(builder.build({
-		tree: sapUiCoreTree,
-		destPath,
-		excludedTasks,
-		includedTasks
-	}).then(() => {
-		return findFiles(expectedPath);
-	}).then((expectedFiles) => {
-		// Check for all directories and files
-		assert.directoryDeepEqual(destPath, expectedPath);
-
-		// Check for all file contents
-		t.deepEqual(expectedFiles.length, 9, "9 files are expected");
-		expectedFiles.forEach((expectedFile) => {
-			const relativeFile = path.relative(expectedPath, expectedFile);
-			const destFile = path.join(destPath, relativeFile);
-			assert.fileEqual(destFile, expectedFile);
-		});
-	}));
-});
-
-const sapUiCoreTree = {
-	"id": "sap.ui.core",
-	"version": "1.0.0",
-	"path": sapUiCorePath,
-	"dependencies": [],
-	"_level": 1,
-	"specVersion": "0.1",
-	"type": "library",
-	"metadata": {
-		"name": "sap.ui.core",
-		"copyright": "Some fancy copyright"
-	},
-	"resources": {
-		"configuration": {
-			"paths": {
-				"src": "main/src",
-				"test": "main/test"
-			}
-		},
-		"pathMappings": {
-			"/resources/": "main/src",
-			"/test-resources/": "main/test"
-		}
-	}
-};
-
-const newLineRegexp = /\r?\n|\r/g;
-
-async function checkFileContentsIgnoreLineFeeds(t, expectedFiles, expectedPath, destPath) {
-	for (let i = 0; i < expectedFiles.length; i++) {
-		const expectedFile = expectedFiles[i];
-		const relativeFile = path.relative(expectedPath, expectedFile);
-		const destFile = path.join(destPath, relativeFile);
-		const currentFileContentPromise = readFile(destFile, "utf8");
-		const expectedFileContentPromise = readFile(expectedFile, "utf8");
-		const assertContents = ([currentContent, expectedContent]) => {
-			if (expectedFile.endsWith(".json")) {
-				try {
-					t.deepEqual(JSON.parse(currentContent), JSON.parse(expectedContent), expectedFile);
-				} catch (e) {
-					t.falsy(e, expectedFile);
-				}
-			}
-			t.is(currentContent.replace(newLineRegexp, "\n"), expectedContent.replace(newLineRegexp, "\n"), relativeFile);
-		};
-		await Promise.all([currentFileContentPromise, expectedFileContentPromise]).then(assertContents);
-	}
-}
